requirements.txt
ccxt==1.79.2
python-dotenv==0.20.0
requests==2.27.1
pandas==1.4.2